Refactor and review in decorator
