+++ src/CarFactory/src/decorator/CarDecorator.java
@@ -1,5 +1,7 @@
 package decorator;
 
+import decorator.utilities.CarDecoratorInterface;
+
 abstract class CarDecorator implements CarDecoratorInterface {
     private CarDecoratorInterface core;
