+++ src/CarFactory/src/decorator/utilities/CarDecoratorInterface.java
@@ -1,4 +1,4 @@
-package decorator;
+package decorator.utilities;
 
 public interface CarDecoratorInterface {
     void decorateCar();
