+++ src/CarFactory/src/decorator/utilities/DecoratedCar.java
@@ -1,9 +1,15 @@
-package decorator;
+package decorator.utilities;
 
 import car.utilities.Car;
 
+import java.util.Set;
+import java.util.TreeSet;
+
 public class DecoratedCar implements CarDecoratorInterface {
+
     private Car car;
+    private Set<String> decorations = new TreeSet<>(); //TODO decorations should be added to this set
+
 
     public DecoratedCar(Car carToDecorate){
         car = carToDecorate;
