+++ src/CarFactory/src/main/Main.java
@@ -1,4 +1,4 @@
-package sample;
+package main;
 
 import javafx.application.Application;
 import javafx.fxml.FXMLLoader;
@@ -18,7 +18,7 @@ public class Main extends Application {
 
     @Override
     public void start(Stage primaryStage) throws Exception{
-        Parent root = FXMLLoader.load(getClass().getResource("sample.fxml"));
+        Parent root = FXMLLoader.load(getClass().getResource("utilities/sample.fxml"));
         primaryStage.setTitle("Car Factory Main Interface");
         primaryStage.setScene(new Scene(root, 700, 500));
         primaryStage.show();
