+++ src/CarFactory/src/main/utilities/sample.fxml
@@ -3,7 +3,7 @@
 
 <?import javafx.scene.control.Button?>
 <?import javafx.scene.control.Label?>
-<GridPane fx:controller="sample.Controller"
+<GridPane fx:controller="main.Controller"
           xmlns:fx="http://javafx.com/fxml" alignment="center" hgap="10" vgap="10">
     <Button text="TestButton" onAction="#buildCar"/>
     <Label GridPane.rowIndex="1" fx:id="test1"/>
