+++ src/CarFactory/test/tests/CarDecoratorTest.java
@@ -5,6 +5,8 @@ import car.utilities.Car;
 import car.utilities.CarProductionRegion;
 import car.utilities.CarType;
 import decorator.*;
+import decorator.utilities.CarDecoratorInterface;
+import decorator.utilities.DecoratedCar;
 import factory.CarFactory;
 import org.junit.Test;
 import org.junit.jupiter.api.Assertions;
@@ -15,7 +17,7 @@ import java.io.PrintStream;
 public class CarDecoratorTest {
 
     @Test
-    public void checkDecorator() {
+    public void shouldDecorateWithSeatsAndClocks() {
         CarFactory carFactory = new CarFactory();
         carFactory.setCarProductionRegion(CarProductionRegion.USA);
         Car car = carFactory.create(CarType.CITY_CAR);
@@ -26,5 +28,23 @@ public class CarDecoratorTest {
         System.setOut(new PrintStream(outContent));
         carDecorations.decorateCar();
         Assertions.assertEquals(car.toString() + " HEATED SEATS DIGITAL CLOCKS", outContent.toString());
+    }
+
+    @Test
+    public void shoulddo() {
+        CarFactory carFactory = new CarFactory();
+        carFactory.setCarProductionRegion(CarProductionRegion.JAPAN);
+        Car car = carFactory.create(CarType.SPORTS_CAR);
+        System.out.println(car);// regular car
+        //CarDecoratorInterface carDecorations = new HeatedSeats(new DecoratedCar(car));//TODO decorateCar() should return DecoratedCar
+        //carDecorations.decorateCar();
+        /*TODO commented part should work
+        DecoratedCar dc = new HeatedSeats(new DecoratedCar(car));
+        */
+        System.out.println("----");
+        System.out.println(car);//TODO this car should be decorated
+
+
+//
     }
 }
