+++ src/CarFactory/test/tests/CarPublisherTest.java
@@ -3,8 +3,8 @@ package tests;
 import car.utilities.Car;
 import car.utilities.CarProductionRegion;
 import car.utilities.CarType;
-import decorator.CarDecoratorInterface;
-import decorator.DecoratedCar;
+import decorator.utilities.CarDecoratorInterface;
+import decorator.utilities.DecoratedCar;
 import decorator.DigitalClocks;
 import decorator.HeatedSeats;
 import factory.CarFactory;
